refactor: split RL trainer into optional verifiers-rl package
