+++ packages/verifiers-rl/README.md
@@ -1,0 +1,18 @@
+# verifiers-rl
+
+Optional RL trainer package for `verifiers`.
+
+Install:
+
+```bash
+uv add verifiers-rl
+```
+
+This package provides:
+
+- `vf-rl`
+- `vf-train`
+- `vf-vllm`
+- `verifiers_rl.rl` (RLTrainer implementation)
+
+`verifiers` core remains usable without this package.
