+++ packages/verifiers-rl/pyproject.toml
@@ -1,0 +1,37 @@
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[project]
+name = "verifiers-rl"
+version = "0.1.0"
+description = "Optional RL trainer package for verifiers"
+readme = "README.md"
+requires-python = ">=3.10,<3.14"
+dependencies = [
+    "verifiers",
+    "torch>=2.8.0,<2.9.0",
+    "transformers>=4.56.2",
+    "accelerate>=1.4.0",
+    "requests",
+    "peft",
+    "wandb",
+    "vllm>=0.10.0,<0.11.0",
+    "liger-kernel>=0.5.10",
+    "deepspeed>=0.17.6",
+    "flash-attn>=2.8.3",
+]
+
+[tool.uv.extra-build-dependencies]
+flash-attn = [{ requirement = "torch", match-runtime = true }]
+
+[tool.uv.extra-build-variables]
+flash-attn = { FLASH_ATTENTION_SKIP_CUDA_BUILD = "TRUE" }
+
+[project.scripts]
+vf-rl = "verifiers_rl.scripts.rl:main"
+vf-train = "verifiers_rl.scripts.train:main"
+vf-vllm = "verifiers_rl.rl.inference.server:main"
+
+[tool.hatch.build.targets.wheel]
+packages = ["verifiers_rl"]
