+++ packages/verifiers-rl/verifiers_rl/__init__.py
@@ -1,0 +1,21 @@
+from verifiers_rl.rl.trainer import (  # noqa: F401
+    GRPOConfig,
+    GRPOTrainer,
+    RLConfig,
+    RLTrainer,
+    get_model,
+    get_model_and_tokenizer,
+    grpo_defaults,
+    lora_defaults,
+)
+
+__all__ = [
+    "get_model",
+    "get_model_and_tokenizer",
+    "RLConfig",
+    "RLTrainer",
+    "GRPOTrainer",
+    "GRPOConfig",
+    "grpo_defaults",
+    "lora_defaults",
+]
