+++ pyproject.toml
@@ -87,25 +87,6 @@ browser = [
     "aiohttp>=3.9.0",
     "python-dotenv>=1.0.0",
 ]
-rl = [
-    "torch>=2.8.0,<2.9.0",
-    "transformers>=4.56.2",
-    "accelerate>=1.4.0",
-    "requests",
-    "peft",
-    "wandb",
-    "vllm>=0.10.0,<0.11.0",
-    "liger-kernel>=0.5.10",
-    "deepspeed>=0.17.6",
-    "flash-attn>=2.8.3",
-]
-
-[tool.uv.extra-build-dependencies]
-flash-attn = [{ requirement = "torch", match-runtime = true }]
-
-[tool.uv.extra-build-variables]
-flash-attn = { FLASH_ATTENTION_SKIP_CUDA_BUILD = "TRUE" }
-
 [project.scripts]
 vf-eval = "verifiers.scripts.eval:main"
 vf-gepa = "verifiers.scripts.gepa:main"
@@ -116,7 +97,7 @@ vf-build = "verifiers.scripts.build:main"
 vf-rl = "verifiers.scripts.rl:main"
 vf-train = "verifiers.scripts.train:main"
 vf-tui = "verifiers.scripts.tui:main"
-vf-vllm = "verifiers.rl.inference.server:main"
+vf-vllm = "verifiers.scripts.vllm:main"
 prime-rl = "verifiers.scripts.prime_rl:main"
 
 # hatchling configuration
