+++ verifiers/__init__.py
@@ -79,14 +79,6 @@ __all__ = [
     "quiet_verifiers",
     "load_environment",
     "print_prompt_completions_sample",
-    "get_model",
-    "get_model_and_tokenizer",
-    "RLTrainer",
-    "RLConfig",
-    "GRPOTrainer",
-    "GRPOConfig",
-    "grpo_defaults",
-    "lora_defaults",
     "cleanup",
     "stop",
     "teardown",
@@ -95,14 +87,14 @@ __all__ = [
 ]
 
 _LAZY_IMPORTS = {
-    "get_model": "verifiers.rl.trainer.utils:get_model",
-    "get_model_and_tokenizer": "verifiers.rl.trainer.utils:get_model_and_tokenizer",
-    "RLConfig": "verifiers.rl.trainer:RLConfig",
-    "RLTrainer": "verifiers.rl.trainer:RLTrainer",
-    "GRPOTrainer": "verifiers.rl.trainer:GRPOTrainer",
-    "GRPOConfig": "verifiers.rl.trainer:GRPOConfig",
-    "grpo_defaults": "verifiers.rl.trainer:grpo_defaults",
-    "lora_defaults": "verifiers.rl.trainer:lora_defaults",
+    "get_model": "verifiers_rl.rl.trainer.utils:get_model",
+    "get_model_and_tokenizer": "verifiers_rl.rl.trainer.utils:get_model_and_tokenizer",
+    "RLConfig": "verifiers_rl.rl.trainer:RLConfig",
+    "RLTrainer": "verifiers_rl.rl.trainer:RLTrainer",
+    "GRPOTrainer": "verifiers_rl.rl.trainer:GRPOTrainer",
+    "GRPOConfig": "verifiers_rl.rl.trainer:GRPOConfig",
+    "grpo_defaults": "verifiers_rl.rl.trainer:grpo_defaults",
+    "lora_defaults": "verifiers_rl.rl.trainer:lora_defaults",
     "MathRubric": "verifiers.rubrics.math_rubric:MathRubric",
     "SandboxEnv": "verifiers.envs.sandbox_env:SandboxEnv",
     "PythonEnv": "verifiers.envs.python_env:PythonEnv",
@@ -124,7 +116,20 @@ def __getattr__(name: str):
     except KeyError:
         raise AttributeError(f"module 'verifiers' has no attribute '{name}'")
     except ModuleNotFoundError as e:
-        # warn that accessed var needs [all] to be installed
+        rl_names = {
+            "get_model",
+            "get_model_and_tokenizer",
+            "RLConfig",
+            "RLTrainer",
+            "GRPOTrainer",
+            "GRPOConfig",
+            "grpo_defaults",
+            "lora_defaults",
+        }
+        if name in rl_names:
+            raise AttributeError(
+                f"To use verifiers.{name}, install as `verifiers-rl`."
+            ) from e
         raise AttributeError(
             f"To use verifiers.{name}, install as `verifiers[all]`. "
         ) from e
@@ -141,16 +146,4 @@ if TYPE_CHECKING:
     from .envs.integrations.textarena_env import TextArenaEnv  # noqa: F401
     from .envs.python_env import PythonEnv  # noqa: F401
     from .envs.sandbox_env import SandboxEnv  # noqa: F401
-    from .rl.trainer import (  # noqa: F401
-        GRPOConfig,
-        GRPOTrainer,
-        RLConfig,
-        RLTrainer,
-        grpo_defaults,
-        lora_defaults,
-    )
-    from .rl.trainer.utils import (  # noqa: F401
-        get_model,
-        get_model_and_tokenizer,
-    )
     from .rubrics.math_rubric import MathRubric  # noqa: F401
