+++ verifiers/scripts/vllm.py
@@ -6,6 +6,6 @@ def main() -> None:
         mod = importlib.import_module("verifiers_rl.rl.inference.server")
     except ModuleNotFoundError as e:
         raise SystemExit(
-            "verifiers.rl.inference.server moved to optional package 'verifiers-rl'. Install with: uv add verifiers-rl"
+            "vf-vllm now lives in the optional 'verifiers-rl' package. Install with: uv add verifiers-rl"
         ) from e
     mod.main()
